Move exceptions into .solr.request
